feat : create PurchaseGiftCardConfirmDialog

+++ features/exploredash/src/main/java/org/dash/wallet/features/exploredash/data/model/purchase/PurchaseGiftCardRequest.kt
@@ -24,5 +24,5 @@ data class PurchaseGiftCardRequest(
     @SerializedName("GiftCardAmount")
     val giftCardAmount: Double? = null,
     @SerializedName("MerchantId")
-    val merchantId: Int? = null
+    val merchantId: Long? = null
 )
